[ingest]: move LeaveConsumerGroupByInstanceID to consumer_group.go
reader_client.go is forked from Loki; adding Tempo-specific helpers
there increases the local diff and complicates future upstream syncs.
Move LeaveConsumerGroupByInstanceID to its own file.
